Dependency Injection Improvements (#79)
* Dependency Injection Improvements

Updated API to fit better with Angular 2 DI.

The main change here is to delay configuring the
redux store until after the app has bootstrapped.
So instead of doing:

```typescript
import { provider } from 'ng2-redux';

bootstrap(App, [
  provider(configureStore()) ]);
```

We instead do:

```typescript
import { NgRedux } from 'ng2-redux';

bootstrap(App, [ NgRedux ]);

@Component({
  // ...
})
class App {
  constructor(private ngRedux: NgRedux) {
    ngRedux.attach(configureStore());
  }
}
```

This is desirable for two reasons:

1. We can now inject angular services into NgRedux
itself
2. We can now inject angular services into redux
middleware and store enhancers.

(1) is needed to resolve the issue with redux
dev tools store updates happening outside the angular
zone.  We need to be able to do an `applicationRef.tick()`
on state update or the Angular UI doesn't respond to
dev tool actions.  Handling it internally this way is
preferable to making the end user remember to hook
it up manually.

(2) is a major feature that is currently impossible
in the old architecture, since previously the middleware
needed to be applied to the store before the Angular DI
was bootstrapped.

I think this approach is much more in line with how
people expect Angular 2 DI to function, and resolves
a number of other thorny init problems.

However it does represent an API change, and thus
will require a major version bump if adopted.

Also in this commit:
- some minor code cleanup
- switch to the more size-efficient rxjs import pattern
- marked the `select` api as deprecated.

* Added DI-injected middleware example with new DI strategy

Also added redux-logger to the example app as an example of using vanilla middlewares from the redux ecosystem.

* Added unit test for double store configuration error.

* Added test for devtools update mechanism.