fix(build): No chai for you
The chai type def has this...

```
interface Object {
    should: Chai.Assertion;
}
```

When we run build, on files that have things like:

`let x = (y:Object):Obect => { }`

In the resulting file.d.ts

```
/// <reference types="chai" />
```

Gets added, causing things to blow up for people. Not good.

Temp fix for now - when running build, uninstall the types - then
re-install them after.

Not happy with this, but rather this than having a patch causing
breaking builds for using 3.3.8.